+++ src/test/java/tk/freaxsoftware/extras/bus/test/annotation/AnnotationUtilTest.java
@@ -19,9 +19,9 @@
 package tk.freaxsoftware.extras.bus.test.annotation;
 
 import org.junit.After;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import spark.utils.Assert;
 import tk.freaxsoftware.extras.bus.MessageBus;
 import tk.freaxsoftware.extras.bus.MessageOptions;
 import tk.freaxsoftware.extras.bus.annotation.AnnotationUtil;
@@ -33,12 +33,15 @@ import tk.freaxsoftware.extras.bus.exceptions.NoSubscriptionMessageException;
  */
 public class AnnotationUtilTest {
     
+    private TestReceiver testReceiver;
+    
     public AnnotationUtilTest() {
     }
     
     @Before
     public void setUp() throws InstantiationException, IllegalAccessException {
-        AnnotationUtil.subscribeReceiverClass(TestReceiver.class);
+        testReceiver = new TestReceiver();
+        AnnotationUtil.subscribeReceiverInstance(testReceiver);
     }
     
     @After
@@ -48,16 +51,18 @@ public class AnnotationUtilTest {
     }
     
     @Test
-    public void testMessage() {
+    public void testMessage() throws Exception {
         MessageBus.fire(TestReceiver.TEST_MESSAGE, null, MessageOptions.Builder.newInstance().deliveryCall().callback((res) -> {
-            Assert.notNull(res.getContent());
+            Assert.assertNotNull(res.getContent());
         }).build());
+        Thread.sleep(1000);
+        Assert.assertTrue(testReceiver.calledPattern);
     }
     
     @Test(expected = NoSubscriptionMessageException.class)
     public void testMessageError() {
         MessageBus.fire(TestReceiver.TEST_MESSAGE_ERROR, null, MessageOptions.Builder.newInstance().deliveryCall().callback((res) -> {
-            Assert.notNull(res.getContent());
+            Assert.assertNotNull(res.getContent());
         }).build());
     }
 }